refactor : creating products & auth the product create.

+++ migrations/sqls/20211108105714-products-table-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name VARCHAR(64) NOT NULL,
-    price integer NOT NULL,
+    price NUMERIC(10,3) NOT NULL,
     category  varchar(64)
 );
